refactoring, unused code deleted
