+++ Cargo.toml
@@ -13,10 +13,8 @@ bb8 =  "0.7"
 bb8-diesel = "0.2.1"
 bb8-postgres = "0.8.0"
 chrono = "0.4.19"
-diesel = { version = "1.4", features = ["r2d2", "postgres", "chrono"], default-features = false }
-dotenv = "0.15.0"
+diesel = { version = "1.4", features = ["postgres", "chrono"], default-features = false }
 futures = "0.3.21"
-lazy_static = "1.4.0"
 pepe-config = "0.0.5"
 reqwest = { version = "0.11.10", default-features = false, features = ["json","rustls-tls"] }
 serde = { version = "1.0.137", features = ["derive"] }
