+++ src/errors.rs
@@ -19,7 +19,7 @@ pub enum ServerError {
     #[error("DB error: {0}")]
     DBError(#[from] diesel::result::Error),
     #[error(transparent)]
-    Other(#[from] anyhow::Error)
+    Other(#[from] anyhow::Error),
 }
 
 impl From<ServerError> for HttpResponse {
@@ -32,7 +32,9 @@ impl From<ServerError> for HttpResponse {
                 HttpResponse::Ok().json(handlers::models::ExecutorResponse::no_target())
             }
             ServerError::PoolError(cause) => HttpResponse::InternalServerError().json(cause),
-            ServerError::DBError(source) => HttpResponse::InternalServerError().json(source.to_string()),
+            ServerError::DBError(source) => {
+                HttpResponse::InternalServerError().json(source.to_string())
+            }
             ServerError::Other(source) => HttpResponse::InternalServerError().json(json!({
                 "reason": source.to_string()
             })),
@@ -41,15 +43,3 @@ impl From<ServerError> for HttpResponse {
 }
 
 impl actix_web::error::ResponseError for ServerError {}
-
-pub fn wrap_err(e: anyhow::Error) -> ServerError {
-    return ServerError::WrappedErr {
-        cause: format!("cause : {}", e),
-    };
-}
-
-pub fn wrap_client_err(e: anyhow::Error) -> ServerError {
-    return ServerError::Unsuccessfull {
-        info: vec![format!("cause : {}", e)],
-    };
-}
